+++ Datalib/src/main/java/com/ing/datalib/component/TestCase.java
@@ -1570,10 +1570,7 @@ public class TestCase extends DataModel {
     @Override
     public Boolean rename(String newName) {
         TestCase existing = getScenario().getTestCaseByName(newName);
-        if (
-            (existing == null || existing == this) &&
-            getScenario().getReusableTestCaseByName(getScenario().getName(), newName) == null
-        ) {
+        if (existing == null || existing == this) {
             if (FileUtils.renameFile(getLocation(), newName + getFormat().extension())) {
                 getProject().refactorTestCase(getScenario().getName(), name, newName);
                 name = newName;
@@ -1584,12 +1581,8 @@ public class TestCase extends DataModel {
     }
 
     public Boolean renameReusable(String newName) {
-        TestCase existingReusable = getScenario()
-            .getReusableTestCaseByName(getScenario().getName(), newName);
-        if (
-            getScenario().getTestCaseByName(getScenario().getName(), newName) == null &&
-            (existingReusable == null || existingReusable == this)
-        ) {
+        TestCase existing = getScenario().getTestCaseByName(newName);
+        if (existing == null || existing == this) {
             if (FileUtils.renameFile(getLocation(), newName + getFormat().extension())) {
                 getProject().refactorTestCase(getScenario().getName(), name, newName);
                 name = newName;
@@ -1605,7 +1598,8 @@ public class TestCase extends DataModel {
      * @return true if successful, false if a test case with the new name already exists
      */
     public Boolean renameSharedReusable(String newName) {
-        if (getScenario().getTestCaseByName(newName) == null) {
+        TestCase existing = getScenario().getTestCaseByName(newName);
+        if (existing == null || existing == this) {
             if (FileUtils.renameFile(getLocation(), newName + getFormat().extension())) {
                 getProject().refactorTestCase(getScenario().getName(), name, newName);
                 name = newName;